Move bootstrapchannelbuilder to its own package
