+++ upup/pkg/fi/cloudup/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "apply_cluster.go",
-        "bootstrapchannelbuilder.go",
         "containerd.go",
         "defaults.go",
         "dns.go",
@@ -26,7 +25,6 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//:go_default_library",
-        "//channels/pkg/api:go_default_library",
         "//cmd/kops-controller/pkg/config:go_default_library",
         "//dns-controller/pkg/dns:go_default_library",
         "//dnsprovider/pkg/dnsprovider:go_default_library",
@@ -44,14 +42,11 @@ go_library(
         "//pkg/client/simple/vfsclientset:go_default_library",
         "//pkg/dns:go_default_library",
         "//pkg/featureflag:go_default_library",
-        "//pkg/kubemanifest:go_default_library",
         "//pkg/model:go_default_library",
         "//pkg/model/alimodel:go_default_library",
         "//pkg/model/awsmodel:go_default_library",
         "//pkg/model/azuremodel:go_default_library",
         "//pkg/model/components:go_default_library",
-        "//pkg/model/components/addonmanifests:go_default_library",
-        "//pkg/model/components/addonmanifests/dnscontroller:go_default_library",
         "//pkg/model/components/etcdmanager:go_default_library",
         "//pkg/model/components/kubeapiserver:go_default_library",
         "//pkg/model/components/node-authorizer:go_default_library",
@@ -71,14 +66,13 @@ go_library(
         "//upup/pkg/fi/cloudup/aliup:go_default_library",
         "//upup/pkg/fi/cloudup/awsup:go_default_library",
         "//upup/pkg/fi/cloudup/azure:go_default_library",
+        "//upup/pkg/fi/cloudup/bootstrapchannelbuilder:go_default_library",
         "//upup/pkg/fi/cloudup/cloudformation:go_default_library",
         "//upup/pkg/fi/cloudup/do:go_default_library",
         "//upup/pkg/fi/cloudup/gce:go_default_library",
         "//upup/pkg/fi/cloudup/openstack:go_default_library",
         "//upup/pkg/fi/cloudup/terraform:go_default_library",
-        "//upup/pkg/fi/fitasks:go_default_library",
         "//upup/pkg/fi/loader:go_default_library",
-        "//upup/pkg/fi/utils:go_default_library",
         "//util/pkg/architectures:go_default_library",
         "//util/pkg/env:go_default_library",
         "//util/pkg/hashing:go_default_library",
@@ -137,6 +131,7 @@ go_test(
         "//upup/models:go_default_library",
         "//upup/pkg/fi:go_default_library",
         "//upup/pkg/fi/cloudup/awsup:go_default_library",
+        "//upup/pkg/fi/cloudup/bootstrapchannelbuilder:go_default_library",
         "//upup/pkg/fi/fitasks:go_default_library",
         "//util/pkg/architectures:go_default_library",
         "//util/pkg/hashing:go_default_library",
