+++ upup/pkg/fi/cloudup/bootstrapchannelbuilder/BUILD.bazel
@@ -1,0 +1,25 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["bootstrapchannelbuilder.go"],
+    importpath = "k8s.io/kops/upup/pkg/fi/cloudup/bootstrapchannelbuilder",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//channels/pkg/api:go_default_library",
+        "//pkg/apis/kops:go_default_library",
+        "//pkg/assets:go_default_library",
+        "//pkg/featureflag:go_default_library",
+        "//pkg/kubemanifest:go_default_library",
+        "//pkg/model:go_default_library",
+        "//pkg/model/components/addonmanifests:go_default_library",
+        "//pkg/model/components/addonmanifests/dnscontroller:go_default_library",
+        "//pkg/model/iam:go_default_library",
+        "//pkg/templates:go_default_library",
+        "//upup/pkg/fi:go_default_library",
+        "//upup/pkg/fi/fitasks:go_default_library",
+        "//upup/pkg/fi/utils:go_default_library",
+        "//vendor/github.com/blang/semver/v4:go_default_library",
+        "//vendor/k8s.io/klog/v2:go_default_library",
+    ],
+)
